requirements.txt
# Core LangChain and LLM
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# UI Framework
streamlit>=1.30.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# NLP and Text Processing
spacy>=3.7.0
nltk>=3.8.0

# Web Scraping and APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.25.0

# Data Manipulation
pandas>=2.0.0

# Vector Store (Optional for semantic search)
chromadb>=0.4.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.6
pillow>=10.0.0
